+++ app/src/main/java/com/ebsolutions/papertrail/financialdataproviderservice/user/UserRepository.java
@@ -1,8 +1,8 @@
 package com.ebsolutions.papertrail.financialdataproviderservice.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import org.springframework.stereotype.Component;
 
-@Repository
+@Component
 public interface UserRepository extends JpaRepository<User, Long> {
 }
