+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/UserSteps.java
@@ -1,34 +1,24 @@
 package com.ebsolutions.papertrail.financialdataproviderservice;
 
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 
-import com.ebsolutions.papertrail.financialdataproviderservice.common.exception.DataProcessingException;
 import com.ebsolutions.papertrail.financialdataproviderservice.config.Constants;
-import com.ebsolutions.papertrail.financialdataproviderservice.model.ServerError;
-import com.ebsolutions.papertrail.financialdataproviderservice.tooling.BaseTest;
+import com.ebsolutions.papertrail.financialdataproviderservice.tooling.FullIntegrationLayerBaseTest;
 import com.ebsolutions.papertrail.financialdataproviderservice.user.User;
-import com.ebsolutions.papertrail.financialdataproviderservice.user.UserRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import java.io.UnsupportedEncodingException;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Assertions;
 import org.springframework.http.HttpStatus;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.MvcResult;
 
-@RequiredArgsConstructor
-public class UserSteps extends BaseTest {
-  protected final UserRepository userRepository;
+public class UserSteps extends FullIntegrationLayerBaseTest {
 
-  protected MvcResult result;
-  protected List<User> users = new ArrayList<>();
+  private MvcResult result;
   private User expectedUserOne;
   private User expectedUserTwo;
 
@@ -49,28 +39,14 @@ public class UserSteps extends BaseTest {
             .firstName("second")
             .lastName("user")
             .build();
-
-    users.add(expectedUserOne);
-    users.add(expectedUserTwo);
-
-    when(userRepository.findAll()).thenReturn(users);
   }
 
-  @And("no users exist")
-  public void noUsersExist() {
-    when(userRepository.findAll()).thenReturn(Collections.emptyList());
-  }
 
   @When("the get all users endpoint is invoked")
   public void theGetAllUsersEndpointIsInvoked() throws Exception {
     result = mockMvc.perform(get(Constants.USERS_URL)).andReturn();
   }
 
-  @Then("the correct empty users response is returned")
-  public void theCorrectEmptyUsersResponseIsReturned() {
-    MockHttpServletResponse mockHttpServletResponse = result.getResponse();
-    Assertions.assertEquals(HttpStatus.NO_CONTENT.value(), mockHttpServletResponse.getStatus());
-  }
 
   @Then("the correct users are returned")
   public void theCorrectUsersAreReturned()
@@ -87,28 +63,6 @@ public class UserSteps extends BaseTest {
 
     User userTwo = users.getLast();
     assertUserDtoAgainstUser(expectedUserTwo, userTwo);
-  }
-
-  @And("the connection to the database fails")
-  public void theConnectionToTheDatabaseFails() {
-    DataProcessingException dataProcessingException =
-        new DataProcessingException("Generic Exception Message!");
-
-    when(userRepository.findAll()).thenThrow(dataProcessingException);
-  }
-
-  @Then("the correct failure response is returned")
-  public void theCorrectFailureResponseIsReturned()
-      throws UnsupportedEncodingException, JsonProcessingException {
-    MockHttpServletResponse mockHttpServletResponse = result.getResponse();
-
-    Assertions.assertEquals(HttpStatus.INTERNAL_SERVER_ERROR.value(),
-        mockHttpServletResponse.getStatus());
-
-    String content = mockHttpServletResponse.getContentAsString();
-
-    ServerError serverError = objectMapper.readValue(content, ServerError.class);
-    Assertions.assertEquals("Generic Exception Message!", serverError.getMessage());
   }
 
   private void assertUserDtoAgainstUser(User expectedUser, User actualUser) {
