+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/tooling/FullIntegrationConfig.java
@@ -1,0 +1,6 @@
+package com.ebsolutions.papertrail.financialdataproviderservice.tooling;
+
+//@CucumberContextConfiguration
+//@ActiveProfiles("full-integration-layer")
+public class FullIntegrationConfig {
+}
