+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/tooling/MockedIntegrationConfig.java
@@ -1,0 +1,6 @@
+package com.ebsolutions.papertrail.financialdataproviderservice.tooling;
+
+//@CucumberContextConfiguration
+//@ActiveProfiles("mocked-integration-layer")
+public class MockedIntegrationConfig {
+}
