+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/tooling/MockedIntegrationLayerBaseTest.java
@@ -1,0 +1,23 @@
+package com.ebsolutions.papertrail.financialdataproviderservice.tooling;
+
+import com.ebsolutions.papertrail.financialdataproviderservice.user.UserRepository;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@ContextConfiguration(classes = MockedIntegrationConfig.class)
+public class MockedIntegrationLayerBaseTest {
+  protected final ObjectMapper objectMapper = new ObjectMapper();
+
+  @Autowired
+  protected MockMvc mockMvc;
+
+  @MockitoBean
+  protected UserRepository userRepository;
+}
